Add swagger docs and validation, correct todo service code, and fix async usage and unit tests

+++ src/todo/todo.service.ts
@@ -1,8 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Todo } from './todo.entity';
-import { TodoDto } from './dto/todo.dto'
+import { Todo } from './model/todo.entity';
+import { TodoDto } from './dto/todo.dto';
 
 @Injectable()
 export class TodoService {
@@ -11,42 +11,37 @@ export class TodoService {
     private todoRepository: Repository<Todo>,
   ) {}
 
-  findAll(): Promise<Todo[]> {
+  async create(todoDto: TodoDto) {
+    const todo = new Todo();
+    todo.title = todoDto.title;
+    todo.notes = todoDto.notes;
+    todo.dueDate = todoDto.dueDate;
+    todo.priority = todoDto.priority;
+    todo.category = todoDto.category;
+    todo.isCompleted = todoDto.isCompleted;
+
+    return this.todoRepository.save(todo);
+  }
+
+  async findAll() {
     return this.todoRepository.find();
   }
 
-  create(todoDto: TodoDto): Promise<Todo> {
-    const todo = new Todo();
-    todo.title = todoDto.title;
-    todo.notes = todoDto.notes;
-    todo.dueDate = todoDto.dueDate;
-    todo.priority = todoDto.priority;
-    todo.category = todoDto.category;
-
-    return this.todoRepository.save(todo);
+  async findOne(id: number) {
+    const todo: Todo = await this.todoRepository.findOneBy({ id });
+    if (!todo) throw new BadRequestException('Todo does not exist.');
+    return todo;
   }
 
-  findOne(id: number): Promise<Todo | null> {
-    return this.todoRepository.findOneBy({ id });
+  async update(id: number, todoDto: Partial<TodoDto>) {
+    const todo: Todo = await this.todoRepository.findOneBy({ id });
+    if (!todo) throw new BadRequestException('Todo does not exist.');
+    return this.todoRepository.save({ ...todo, ...todoDto });
   }
 
-  update(id: number, todoDto: TodoDto): Promise<Todo> {
-    const todo = new Todo();
-    todo.id = id;
-    todo.title = todoDto.title;
-    todo.notes = todoDto.notes;
-    todo.dueDate = todoDto.dueDate;
-    todo.priority = todoDto.priority;
-    todo.category = todoDto.category;
-
-    return this.todoRepository.save(todo);
-  }
-
-  async remove(id: number): Promise<void> {
+  async remove(id: number) {
+    if (!(await this.todoRepository.findOneBy({ id })))
+      throw new BadRequestException('Todo does not exist.');
     await this.todoRepository.delete(id);
-  }
-
-  async complete(id: number): Promise<void> {
-    await this.todoRepository.update(id, { isCompleted: true })
   }
 }
